+++ lib/src/document/relationship.dart
@@ -2,10 +2,10 @@ import 'dart:async';
 
 import 'package:json_api/src/client/client.dart';
 import 'package:json_api/src/document/document.dart';
+import 'package:json_api/src/document/identifier.dart';
 import 'package:json_api/src/document/identifier_object.dart';
 import 'package:json_api/src/document/link.dart';
 import 'package:json_api/src/document/resource_object.dart';
-import 'package:json_api/src/identifier.dart';
 import 'package:json_api/src/nullable.dart';
 
 /// A relationship. Can be to-one or to-many.